requirements.txt
torch
transformers
streamlit
Pillow
accelerate
sentence-transformers
python-dotenv
huggingface-hub
groq
streamlit-chat
